Extract shell from UI
